+++ projects/plugins/crm/src/automation/commons/triggers/contacts/class-contact-created.php
@@ -1,0 +1,85 @@
+<?php
+/**
+ * Jetpack CRM Automation Contact_Created trigger.
+ *
+ * @package automattic/jetpack-crm
+ * @since $$next-version$$
+ */
+
+namespace Automattic\Jetpack\CRM\Automation\Triggers;
+
+use Automattic\Jetpack\CRM\Automation\Automation_Workflow;
+use Automattic\Jetpack\CRM\Automation\Base_Trigger;
+
+/**
+ * Adds the Contact_Created class.
+ *
+ * @since $$next-version$$
+ */
+class Contact_Created extends Base_Trigger {
+
+	/**
+	 * The Automation workflow object.
+	 *
+	 * @since $$next-version$$
+	 * @var Automation_Workflow
+	 */
+	protected $workflow;
+
+	/**
+	 * Get the slug name of the trigger.
+	 *
+	 * @since $$next-version$$
+	 *
+	 * @return string The slug name of the trigger.
+	 */
+	public static function get_slug(): string {
+		return 'jpcrm/contact_created';
+	}
+
+	/**
+	 * Get the title of the trigger.
+	 *
+	 * @since $$next-version$$
+	 *
+	 * @return string|null The title of the trigger.
+	 */
+	public static function get_title(): ?string {
+		return __( 'New Contact', 'zero-bs-crm' );
+	}
+
+	/**
+	 * Get the description of the trigger.
+	 *
+	 * @since $$next-version$$
+	 *
+	 * @return string|null The description of the trigger.
+	 */
+	public static function get_description(): ?string {
+		return __( 'Triggered when a CRM contact is added', 'zero-bs-crm' );
+	}
+
+	/**
+	 * Get the category of the trigger.
+	 *
+	 * @since $$next-version$$
+	 *
+	 * @return string|null The category of the trigger.
+	 */
+	public static function get_category(): ?string {
+		return __( 'contact', 'zero-bs-crm' );
+	}
+
+	/**
+	 * Listen to the desired event.
+	 *
+	 * @since $$next-version$$
+	 */
+	protected function listen_to_event() {
+		add_action(
+			'jpcrm_contact_created',
+			array( $this, 'execute_workflow' )
+		);
+	}
+
+}