+++ projects/plugins/crm/tests/php/automation/contacts/class-contact-trigger-test.php
@@ -2,12 +2,11 @@
 
 namespace Automattic\Jetpack\CRM\Automation\Tests;
 
-use Automattic\Jetpack\CRM\Automation\Automation_Engine;
 use Automattic\Jetpack\CRM\Automation\Automation_Workflow;
 use Automattic\Jetpack\CRM\Automation\Triggers\Contact_Before_Deleted;
+use Automattic\Jetpack\CRM\Automation\Triggers\Contact_Created;
 use Automattic\Jetpack\CRM\Automation\Triggers\Contact_Deleted;
 use Automattic\Jetpack\CRM\Automation\Triggers\Contact_Email_Updated;
-use Automattic\Jetpack\CRM\Automation\Triggers\Contact_New;
 use Automattic\Jetpack\CRM\Automation\Triggers\Contact_Status_Updated;
 use Automattic\Jetpack\CRM\Automation\Triggers\Contact_Updated;
 use WorDBless\BaseTestCase;
@@ -40,7 +39,7 @@ class Contact_Trigger_Test extends BaseTestCase {
 
 		// Build a PHPUnit mock Automation_Workflow
 		$workflow = $this->getMockBuilder( Automation_Workflow::class )
-			->setConstructorArgs( array( $workflow_data, new Automation_Engine() ) )
+			->setConstructorArgs( array( $workflow_data ) )
 			->onlyMethods( array( 'execute' ) )
 			->getMock();
 
@@ -59,7 +58,7 @@ class Contact_Trigger_Test extends BaseTestCase {
 		);
 
 		// Run the contact_update action.
-		do_action( 'jpcrm_automation_contact_update', $contact_data );
+		do_action( 'jpcrm_contact_updated', $contact_data );
 	}
 
 	/**
@@ -73,7 +72,7 @@ class Contact_Trigger_Test extends BaseTestCase {
 
 		// Build a PHPUnit mock Automation_Workflow
 		$workflow = $this->getMockBuilder( Automation_Workflow::class )
-			->setConstructorArgs( array( $workflow_data, new Automation_Engine() ) )
+			->setConstructorArgs( array( $workflow_data ) )
 			->onlyMethods( array( 'execute' ) )
 			->getMock();
 
@@ -92,31 +91,31 @@ class Contact_Trigger_Test extends BaseTestCase {
 		);
 
 		// Run the contact_status_update action.
-		do_action( 'jpcrm_automation_contact_status_update', $contact_data );
+		do_action( 'jpcrm_contact_status_updated', $contact_data );
 	}
 
 	/**
 	 * @testdox Test the contact new trigger executes the workflow with an action
 	 */
-	public function test_contact_new_trigger() {
+	public function test_contact_created_trigger() {
 
-		$workflow_data = $this->automation_faker->workflow_without_initial_step_customize_trigger( 'jpcrm/contact_new' );
+		$workflow_data = $this->automation_faker->workflow_without_initial_step_customize_trigger( 'jpcrm/contact_created' );
 
-		$trigger = new Contact_New();
+		$trigger = new Contact_Created();
 
 		// Build a PHPUnit mock Automation_Workflow
 		$workflow = $this->getMockBuilder( Automation_Workflow::class )
-			->setConstructorArgs( array( $workflow_data, new Automation_Engine() ) )
+			->setConstructorArgs( array( $workflow_data ) )
 			->onlyMethods( array( 'execute' ) )
 			->getMock();
 
-		// Init the Contact_New trigger.
+		// Init the Contact_Created trigger.
 		$trigger->init( $workflow );
 
 		// Fake event data.
 		$contact_data = $this->automation_faker->contact_data();
 
-		// We expect the workflow to be executed on contact_new event with the contact data
+		// We expect the workflow to be executed on contact_created event with the contact data
 		$workflow->expects( $this->once() )
 		->method( 'execute' )
 		->with(
@@ -124,8 +123,8 @@ class Contact_Trigger_Test extends BaseTestCase {
 			$this->equalTo( $contact_data )
 		);
 
-		// Run the contact_new action.
-		do_action( 'jpcrm_automation_contact_new', $contact_data );
+		// Run the contact_created action.
+		do_action( 'jpcrm_contact_created', $contact_data );
 	}
 
 	/**
@@ -139,7 +138,7 @@ class Contact_Trigger_Test extends BaseTestCase {
 
 		// Build a PHPUnit mock Automation_Workflow
 		$workflow = $this->getMockBuilder( Automation_Workflow::class )
-			->setConstructorArgs( array( $workflow_data, new Automation_Engine() ) )
+			->setConstructorArgs( array( $workflow_data ) )
 			->onlyMethods( array( 'execute' ) )
 			->getMock();
 
@@ -158,7 +157,7 @@ class Contact_Trigger_Test extends BaseTestCase {
 		);
 
 		// Run the contact_email_update action.
-		do_action( 'jpcrm_automation_contact_email_update', $contact_data );
+		do_action( 'jpcrm_contact_email_updated', $contact_data );
 	}
 
 	/**
@@ -172,7 +171,7 @@ class Contact_Trigger_Test extends BaseTestCase {
 
 		// Build a PHPUnit mock Automation_Workflow
 		$workflow = $this->getMockBuilder( Automation_Workflow::class )
-			->setConstructorArgs( array( $workflow_data, new Automation_Engine() ) )
+			->setConstructorArgs( array( $workflow_data ) )
 			->onlyMethods( array( 'execute' ) )
 			->getMock();
 
@@ -191,7 +190,7 @@ class Contact_Trigger_Test extends BaseTestCase {
 		);
 
 		// Run the contact_deleted action.
-		do_action( 'jpcrm_automation_contact_delete', $contact_data );
+		do_action( 'jpcrm_contact_deleted', $contact_data );
 	}
 
 	/**
@@ -205,7 +204,7 @@ class Contact_Trigger_Test extends BaseTestCase {
 
 		// Build a PHPUnit mock Automation_Workflow
 		$workflow = $this->getMockBuilder( Automation_Workflow::class )
-			->setConstructorArgs( array( $workflow_data, new Automation_Engine() ) )
+			->setConstructorArgs( array( $workflow_data ) )
 			->onlyMethods( array( 'execute' ) )
 			->getMock();
 
@@ -224,7 +223,7 @@ class Contact_Trigger_Test extends BaseTestCase {
 		);
 
 		// Run the contact_before_deleted action.
-		do_action( 'jpcrm_automation_contact_before_delete', $contact_data );
+		do_action( 'jpcrm_contact_before_deleted', $contact_data );
 	}
 
 }