+++ projects/plugins/jetpack/extensions/plugins/ai-assistant-plugin/components/proofread/index.tsx
@@ -1,22 +1,32 @@
 /**
  * External dependencies
  */
+import { aiAssistantIcon, useAiSuggestions } from '@automattic/jetpack-ai-client';
+import { useAnalytics } from '@automattic/jetpack-shared-extension-utils';
+import { serialize } from '@wordpress/blocks';
 import { Modal, Button, Spinner } from '@wordpress/components';
+import { useSelect } from '@wordpress/data';
 import { useState } from '@wordpress/element';
 import { __ } from '@wordpress/i18n';
 import { Icon, close } from '@wordpress/icons';
+import TurndownService from 'turndown';
 /**
  * Internal dependencies
  */
-import useSuggestionsFromAI from '../../../../blocks/ai-assistant/hooks/use-suggestions-from-ai';
-import aiAssistantIcon from '../../../../blocks/ai-assistant/icons/ai-assistant';
 import {
 	delimiter,
 	getDelimitedContent,
 	getInitialSystemPrompt,
 } from '../../../../blocks/ai-assistant/lib/prompt';
-import { getContentFromBlocks } from '../../../../blocks/ai-assistant/lib/utils/block-content';
 import './style.scss';
+
+// Turndown instance
+const turndownService = new TurndownService();
+
+const usePostContent = () => {
+	const blocks = useSelect( select => select( 'core/block-editor' ).getBlocks(), [] );
+	return blocks?.length ? turndownService.turndown( serialize( blocks ) ) : '';
+};
 
 const ModalHeader = ( { loading, onClose } ) => {
 	return (
@@ -32,9 +42,19 @@ const ModalHeader = ( { loading, onClose } ) => {
 	);
 };
 
-export default function Proofread() {
+export default function Proofread( {
+	disabled = false,
+	busy = false,
+}: {
+	disabled?: boolean;
+	busy?: boolean;
+} ) {
 	const [ isProofreadModalVisible, setIsProofreadModalVisible ] = useState( false );
 	const [ suggestion, setSuggestion ] = useState( null );
+	const { tracks } = useAnalytics();
+
+	const postId = useSelect( select => select( 'core/editor' ).getCurrentPostId(), [] );
+	const postContent = usePostContent();
 
 	const toggleProofreadModal = () => {
 		setIsProofreadModalVisible( ! isProofreadModalVisible );
@@ -56,8 +76,10 @@ export default function Proofread() {
 		// TODO: Handle Done
 	};
 
-	const { request, requestingState } = useSuggestionsFromAI( {
-		autoRequest: false,
+	const { request, requestingState } = useAiSuggestions( {
+		askQuestionOptions: {
+			postId,
+		},
 		onSuggestion: handleSuggestion,
 		onDone: handleDone,
 		onError: handleSuggestionError,
@@ -67,19 +89,18 @@ export default function Proofread() {
 		const messages = [
 			getInitialSystemPrompt( {
 				context:
-					'You are an advanced polyglot ghostwriter. Your task is to review blog content, and provide reasonable actions and feedback about the content, without suggesting to rewrite or help with. This functionality is integrated into the Jetpack product developed by Automattic. Users interact with you through a Gutenberg sidebar, you are inside the WordPress editor',
+					'You are an advanced polyglot ghostwriter. Your task is to review blog content and provide reasonable actions and feedback about the content, without suggesting to rewrite it. This functionality is integrated into the Jetpack product developed by Automattic. Users interact with you through a Gutenberg sidebar. You are inside the WordPress editor',
 				rules: [
-					'Format your response in plain text only including break lines.',
-					'Focus on feeeback and not summarize the content.',
-					"Be concise and direct, doesn't repeat the content, and avoid repeat the request. Ex. 'The content delimited ...'",
-					'Do not ask to assist with something more',
+					'Format your response in plain text only including line breaks',
+					'Focus on feedback and do not summarize the content',
+					"Be concise and direct, don't repeat the content and avoid repeating the request. Ex. 'The content delimited ...'",
+					'Do not ask to assist with anything else',
 					'Answer in the same language as the content',
 				],
 				useGutenbergSyntax: false,
 				useMarkdown: false,
 			} ),
 		];
-		const postContent = getContentFromBlocks();
 		messages.push( {
 			role: 'user',
 			content: `Provide a short feedback about the post content delimited with ${ delimiter }. If it could be improved, provide a list of actions on how to do it. ${ getDelimitedContent(
@@ -89,6 +110,9 @@ export default function Proofread() {
 
 		request( messages );
 		toggleProofreadModal();
+		tracks.recordEvent( 'jetpack_ai_get_feedback', {
+			post_id: postId,
+		} );
 	};
 
 	return (
@@ -111,7 +135,12 @@ export default function Proofread() {
 					'jetpack'
 				) }
 			</p>
-			<Button onClick={ handleRequest } variant="secondary">
+			<Button
+				onClick={ handleRequest }
+				variant="secondary"
+				disabled={ ! postContent || disabled }
+				isBusy={ busy }
+			>
 				{ __( 'Generate feedback', 'jetpack' ) }
 			</Button>
 		</div>